+++ src/main/java/com/byb/bhojan/dao/MessPaymentVoucherDao.java
@@ -16,5 +16,5 @@ public interface MessPaymentVoucherDao {
 
 	public void deleteVoucher(MessPaymentVoucher voucher);
 
-	public int getTotalVoucherCount();
+	public long getTotalVoucherCount();
 }
